\name{print.pyRAD.loci}
\alias{print.pyRAD.loci}
\title{Print method for \code{pyRAD.loci} objects}
\description{Prints very basic summary data for a \code{pyRAD.loci} object 
			 rather than streaming the entire thing}
\usage{
print.pyRAD.loci <- function (x, ...)
}
\arguments{
  \item{x}{an object of class \code{pyRAD.loci}}
  \item{dots}{unused}
}
\value{No value. Called for the side-effect of printing attractively to the screen}
\author{Andrew Hipp}
\seealso{
  \code{\link{read.pyRAD}}
}