\name{RADami-package}
\alias{RADami-package}
\alias{RADami}
\docType{package}
\title{Downstream analysis of RADseq phylogenetic data, closely integrated with pyRAD output (http://www.dereneaton.com/software/)}
\description{This package provides tools for manipulating and analyzing RADseq phylogenetic data after clustering, and in many cases after intial
			 phylogenetic analyses have been performed. It begins with output from the pyRAD package by Deren Eataon (Eaton and Ree, 2013) and
			 implements many analyses conducted in a recent phylogenetic analysis of oaks (Hipp et al., accepted).}
\details{
\tabular{ll}{
Package: \tab RADami\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-27\cr
License: \tab GPL-3\cr
}
}
\author{Andrew Hipp

Maintainer: Andrew Hipp <ahipp@mortonarb.org>
}
\references{
Eaton D.A.R. & Ree R.H. (2013). Inferring Phylogeny and Introgression using RADseq Data: An Example from Flowering Plants (Pedicularis: Orobanchaceae). Systematic Biolology 62: 689-706.
Hipp A.L., Eaton D.A.R., Cavender-Bares J., Fitzek E., Nipper R. & Manos P.S. (Accepted pending revision). A framework phylogeny of the American oak clade based on sequenced RAD data. PLoS ONE.
http://www.dereneaton.com/software/
}
\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
